+++ CHANGELOG.md
@@ -12,21 +12,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Makefile` at the project root added for development operations command
 - Generate `stubfiles` from `make stubgen` command
 - `make clean` command for cleaning clickgen cache
+- `make dev` command for development purpose
+- `make docs_gen` command for generating docs
 - Build `xcursorgen` with extra flags
 - `xcursorgen.c` formated with tool **[indent](https://www.gnu.org/software/indent/)**
-- `make dev` command for development purpose
 - Linting & typing fixes inside `clickgen.builders`
 - `Linting`, `pip package caching`, and `stubgen` commands inside [workflows/app-publish.yml](./.github/workflows/app-publish.yml)
 - `ConfigFrame` typing added inside `WindowsCursor` class
+- `clickgen.builder` module docs init
+- docstring `param type` and `rtype` typing with **"or"** inside `Optional` and `Union`
 
 ### Changed
 
+- clean `xcursorgen` build cache automatically on `make` command
 - `CI` pip caching system key changed to `setup.py`
 - Proper typing inheritation inside `clickgen/core.pyi`
 - Linting & Typing fixed in `XCursor` Class `clickgen/builder.py`
 - `xcursorgen/makefile` renamed to `xcursorgen/Makefile`
 - WindowsCursor support `options` instead of `args`
 - clickgen pip dependencies _installation_ method changed inside [workflows/app-ci.yml](./.github/workflows/app-ci.yml)
+- Only `python3` syntax (removed `(object) inheritation`)
 
 ## [1.1.9] - 22 Mar 2021